Zero-warning build.
* Re-save all XIB files and update deployment target to match project.
* Fix some content-size warnings in Log.xib.
* Run Xcode's recommended "project settings update."
* Change some NSUInteger/NSInteger to unsigned/long types.

+++ Source/BSLineNumberRulerView.h
@@ -41,6 +41,6 @@
 
 // Returns the line number (1-based) at the given point. |point| should be in
 // the receiver's coordinate system.
-- (NSUInteger)lineNumberAtPoint:(NSPoint)point;
+- (unsigned long)lineNumberAtPoint:(NSPoint)point;
 
 @end
